use numpy type in base attr access test

+++ tests/data/mypy/correct/attr_access_builtin.py
@@ -5,7 +5,7 @@ from numpydantic import NDArray, Shape
 
 
 class GroundTruth(BaseModel):
-    A: NDArray[Shape["* unit, * height, * width"], np.float64]
+    A: NDArray[Shape["* unit, * height, * width"], float]
 
 
 def use(gt: GroundTruth) -> tuple[int, ...]:
